Moved Provider interfaces into core library
This will allow multiple implementations of the provider to
exist without having everyone depending on rxlifecycle-components
