+++ rxlifecycle-components/src/test/java/com/trello/rxlifecycle/components/RxFragmentLifecycleTest.java
@@ -16,6 +16,7 @@ package com.trello.rxlifecycle.components;
 
 import android.app.Fragment;
 import com.trello.rxlifecycle.FragmentEvent;
+import com.trello.rxlifecycle.FragmentLifecycleProvider;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
