+++ rxlifecycle-components/src/test/java/com/trello/rxlifecycle/components/support/RxSupportFragmentLifecycleTest.java
@@ -17,7 +17,7 @@ package com.trello.rxlifecycle.components.support;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentActivity;
 import com.trello.rxlifecycle.FragmentEvent;
-import com.trello.rxlifecycle.components.FragmentLifecycleProvider;
+import com.trello.rxlifecycle.FragmentLifecycleProvider;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
